+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/logging/LoggingAsserter.java
@@ -1,9 +1,8 @@
-package sa.com.cloudsolutions.antikythera.evaluator;
+package sa.com.cloudsolutions.antikythera.evaluator.logging;
 
 import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.expr.MethodCallExpr;
 import com.github.javaparser.ast.expr.StringLiteralExpr;
-import java.util.List;
 
 /**
  * Provides methods for asserting logging behavior in tests.
